Refactor: 환불 완료시 MyBook에서 삭제되도록 로직 추가

+++ 4Week_Misson/src/main/java/com/ll/exam/sen_books/app/cash/service/CashService.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 public class CashService {
     private final CashLogRepository cashLogRepository;
 
-    public CashLog addCash(Member member, long price, String eventType) {
+    public CashLog addCash(Member member, int price, String eventType) {
         CashLog cashLog = CashLog.builder()
                 .member(member)
                 .price(price)
